requirements.txt
customtkinter==5.2.2
pynput==1.7.7
pyautoit==0.6.5
